Hacer configurables los niveles de clan y pasar exp a Int32
Cambia MAX_LEVEL_GUILD de constante en tiempo de compilación a
variable cargada en runtime, permitiendo configurar dinámicamente
la cantidad de niveles de clan desde Clanes.dat (nueva sección
[GUILDLEVELS], clave MaxGuildLevel) sin necesidad de recompilar.
Los arrays asociados (ExpLevelUpGuild, MembersByLevel,
PriceAcceptMemberGuild) pasan de tamaño fijo a dinámicos, con
ReDim en LoadGuildsConfig apenas se lee MaxGuildLevel.

Convierte los valores de experiencia de clan de Integer (Int16) a
Long (Int32) en todo el código relacionado, para soportar valores de
experiencia requerida por encima de 32767 y evitar overflow. Esto
afecta:
- SendGuildNews: ExpAcu/ExpNe de Integer a Long.
- WriteGuildNews y WriteGuildLeaderInfo: los parámetros ExpActual/
  ExpNecesaria (o ExpAcu/ExpNe) pasan de Integer a Long, y se
  actualiza el protocolo para escribirlos con WriteInt32 en vez de
  WriteInt16.

El overflow ocurría al calcular la experiencia requerida para clanes
en niveles altos (ej. GuildExpLevel7=48000, que supera el máximo de
un Integer), lo que hacía que el panel de clan no se abriera para
fundadores en esos niveles: la asignación fallaba con Overflow antes
de poder mandar el paquete correspondiente al cliente.

Relacionado con el PR de cliente que actualiza HandleGuildNews y
HandleGuildLeaderInfo para leer con ReadInt32 en vez de ReadInt16,
reemplaza los Select Case hardcodeados a 7 niveles por lógica
dinámica basada en los umbrales de recompensa configurados, y hace
dinámico el array cfgGuildMembersByLevel. También relacionado con el
PR de recursos que agrega el nivel 8 de clan a Clanes.dat. Los tres
PRs deben mergearse juntos: el formato del paquete cambia (Int16 →
Int32), así que server y cliente tienen que actualizarse a la vez
para no desincronizar el parser, y el server espera que existan las
claves de nivel 8 en el .dat.

+++ Codigo/Protocol_Writes.bas
@@ -2194,8 +2194,8 @@ Public Sub WriteGuildNews(ByVal UserIndex As Integer, _
                           ByRef guildList() As String, _
                           ByRef MemberList() As Long, _
                           ByVal ClanNivel As Byte, _
-                          ByVal ExpAcu As Integer, _
-                          ByVal ExpNe As Integer)
+                          ByVal ExpAcu As Long, _
+                          ByVal ExpNe As Long)
     On Error GoTo WriteGuildNews_Err
     Dim i   As Long
     Dim Tmp As String
@@ -2215,8 +2215,8 @@ Public Sub WriteGuildNews(ByVal UserIndex As Integer, _
     If Len(Tmp) Then Tmp = Left$(Tmp, Len(Tmp) - 1)
     Call Writer.WriteString8(Tmp)
     Call Writer.WriteInt8(ClanNivel)
-    Call Writer.WriteInt16(ExpAcu)
-    Call Writer.WriteInt16(ExpNe)
+    Call Writer.WriteInt32(ExpAcu)
+    Call Writer.WriteInt32(ExpNe)
     Call modSendData.SendData(ToIndex, UserIndex)
     Exit Sub
 WriteGuildNews_Err:
@@ -2349,8 +2349,8 @@ Public Sub WriteGuildLeaderInfo(ByVal UserIndex As Integer, _
                                 ByVal guildNews As String, _
                                 ByRef joinRequests() As String, _
                                 ByVal NivelDeClan As Byte, _
-                                ByVal ExpActual As Integer, _
-                                ByVal ExpNecesaria As Integer)
+                                ByVal ExpActual As Long, _
+                                ByVal ExpNecesaria As Long)
     On Error GoTo WriteGuildLeaderInfo_Err
     Dim i   As Long
     Dim Tmp As String
@@ -2378,8 +2378,8 @@ Public Sub WriteGuildLeaderInfo(ByVal UserIndex As Integer, _
     If Len(Tmp) Then Tmp = Left$(Tmp, Len(Tmp) - 1)
     Call Writer.WriteString8(Tmp)
     Call Writer.WriteInt8(NivelDeClan)
-    Call Writer.WriteInt16(ExpActual)
-    Call Writer.WriteInt16(ExpNecesaria)
+    Call Writer.WriteInt32(ExpActual)
+    Call Writer.WriteInt32(ExpNecesaria)
     Call modSendData.SendData(ToIndex, UserIndex)
     Exit Sub
 WriteGuildLeaderInfo_Err:
